Fixed configuration of analysis-card
Envsubst can be used on final files (e.g., `demo/index.html`) but should
not be used on files that are meant to be reused (such as polymer
definition files).
This commit adds the ES and LUIGI configuration parameters to the
components so they acn be reused.

It also changes the ES endpoint to the global sefarad
endpoint (sefarad/sefarad) and adds building the image to the CI
pipeline.
This last step is likely to fail, tho, as it is untested.